+++ src/test/scala/com/codecommit/antixml/StAXSpecs.scala
@@ -35,7 +35,7 @@ class StAXSpecs extends Specification {
   
   "StAXParser" should {
     "parse a StreamSource and generate an Elem" in {
-      StAXParser.fromString("<a:a xmlns:a='a'>hi<b attr='value' /> there</a:a>") mustEqual Elem(Some("a"), "a", Map.empty, Group(Text("hi"), Elem(None, "b", Map("attr" -> "value"), Group()), Text(" there")))
+      StAXParser.fromString("<a:a xmlns:a='a'>hi<b attr='value' /> there</a:a>") mustEqual Elem(Some("a"), "a", Attributes(), Group(Text("hi"), Elem(None, "b", Attributes("attr" -> "value"), Group()), Text(" there")))
     }
   }
 }