+++ src/test/scala/com/codecommit/antixml/XMLSpecs.scala
@@ -68,5 +68,5 @@ class XMLSpecs extends Specification {
     }
   }
   
-  def elem(name: String, children: Node*) = Elem(None, name, Map(), Group(children: _*))
+  def elem(name: String, children: Node*) = Elem(None, name, Attributes(), Group(children: _*))
 }