+++ src/test/scala/com/codecommit/antixml/ZipperSpecs.scala
@@ -45,22 +45,22 @@ class ZipperSpecs extends Specification {
   "zipper updates within '\\' results" should {
     "rebuild updated at one level" in {
       val books = bookstore \ "book"
-      val book0 = books(0).copy(attrs=Map("updated" -> "yes"))
-      val book2 = books(2).copy(attrs=Map("updated" -> "yes"))
+      val book0 = books(0).copy(attrs=Attributes("updated" -> "yes"))
+      val book2 = books(2).copy(attrs=Attributes("updated" -> "yes"))
       
       val bookstore2: Group[Node] = books.updated(0, book0).updated(2, book2).unselect    // ensure we have NodeSeq
       
       // find afresh without using \
-      bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
-      bookstore2.head.asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Map()
-      bookstore2.head.asInstanceOf[Elem].children(2).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
+      bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
+      bookstore2.head.asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Attributes()
+      bookstore2.head.asInstanceOf[Elem].children(2).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
     }
     
     "rebuild updated at two levels" in {
       val authors = bookstore \ "book" \ "author"
-      val author0 = authors(0).copy(attrs=Map("updated" -> "yes"))
-      val author2 = authors(2).copy(attrs=Map("updated" -> "yes"))
-      val author3 = authors(3).copy(attrs=Map("updated" -> "yes"))
+      val author0 = authors(0).copy(attrs=Attributes("updated" -> "yes"))
+      val author2 = authors(2).copy(attrs=Attributes("updated" -> "yes"))
+      val author3 = authors(3).copy(attrs=Attributes("updated" -> "yes"))
       
       val bookstore2: Group[Node] = authors.updated(0, author0).updated(2, author2).updated(3, author3).unselect.unselect
       
@@ -69,34 +69,34 @@ class ZipperSpecs extends Specification {
       bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].name mustEqual "book"
       
       bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].children must haveSize(2)
-      bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
+      bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
       
       bookstore2.head.asInstanceOf[Elem].children(1).asInstanceOf[Elem].children must haveSize(2)
-      bookstore2.head.asInstanceOf[Elem].children(1).asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Map()
+      bookstore2.head.asInstanceOf[Elem].children(1).asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Attributes()
       
       bookstore2.head.asInstanceOf[Elem].children(2).asInstanceOf[Elem].children must haveSize(3)
-      bookstore2.head.asInstanceOf[Elem].children(2).asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
-      bookstore2.head.asInstanceOf[Elem].children(2).asInstanceOf[Elem].children(2).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
+      bookstore2.head.asInstanceOf[Elem].children(2).asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
+      bookstore2.head.asInstanceOf[Elem].children(2).asInstanceOf[Elem].children(2).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
     }
     
     "rebuild after a map at the first level" in {
       val books = bookstore \ "book"
-      val books2 = books map { _.copy(attrs=Map("updated" -> "yes")) }
+      val books2 = books map { _.copy(attrs=Attributes("updated" -> "yes")) }
       
       val bookstore2: Group[Node] = books2.unselect
       bookstore2.head.asInstanceOf[Elem].children must haveSize(3)
       
-      // find afresh without using \
-      bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
-      bookstore2.head.asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
-      bookstore2.head.asInstanceOf[Elem].children(2).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
+      // find afresh without using \                                                    
+      bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
+      bookstore2.head.asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
+      bookstore2.head.asInstanceOf[Elem].children(2).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
     }
     
     "rebuild after a flatMap at the first level" in {
       val books = bookstore \ "book"
       val books2 = books flatMap { 
         case e @ Elem(_, _, _, children) if children.length > 2 =>
-          List(e.copy(attrs=Map("updated" -> "yes")), e.copy(attrs=Map("updated" -> "yes")))
+          List(e.copy(attrs=Attributes("updated" -> "yes")), e.copy(attrs=Attributes("updated" -> "yes")))
         
         case _ => Nil
       }
@@ -105,8 +105,8 @@ class ZipperSpecs extends Specification {
       bookstore2.head.asInstanceOf[Elem].children must haveSize(2)
       
       // find afresh without using \
-      bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
-      bookstore2.head.asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Map("updated" -> "yes")
+      bookstore2.head.asInstanceOf[Elem].children(0).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
+      bookstore2.head.asInstanceOf[Elem].children(1).asInstanceOf[Elem].attrs mustEqual Attributes("updated" -> "yes")
       
       (bookstore2 \ "book" \ "title" \ *) must beLike((_:Node) match { case Text("Programming Scala") => ok }).forall
     }